+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 - none
 
 # v2023.11.1-beta
+- ce - Add shell-function shCheatengineUpdate().
 - ci - Create github-action to autobuild cheat-engine.